telegram-bot/requirements.txt
python-telegram-bot==21.0.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic-settings==2.6.0
structlog==24.4.0
fastapi==0.115.0
uvicorn[standard]==0.30.0

